Django/backend/requirements.txt
asgiref==3.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
cloudinary==1.44.1
Django==5.2.8
django-cors-headers==4.9.0
django-environ==0.12.0
djangorestframework==3.16.1
gunicorn==23.0.0
idna==3.11
mysqlclient==2.2.7
packaging==25.0
paytmchecksum==1.7.0
PyJWT==2.10.1
python-dotenv==1.2.1
razorpay==2.0.0
requests==2.32.5
six==1.17.0
sqlparse==0.5.4
stripe==14.0.1
typing_extensions==4.15.0
urllib3==2.5.0
whitenoise==6.11.0
